requirements.txt
# ===============================================
# Note to avoid future compatibility issues: 
# (1) Use a virtual environment.
# (2) Please add version numbers when adding new packages.
# ===============================================

# General Utils
numpy>=1.26,<2

# General Video Processing
opencv-python>=4.7 #OpenCV for video processing
scenedetect>=0.6 # PySceneDetect

# Light Captioning - BLIP 
torch == 2.9.1 
transformers==4.57.1
Pillow== 12.0.0

# Heavy Captioning - InternLM-XComposer2
sentencepiece==0.2.1
torchvision==0.24.1
einops==0.8.1

# --- Audio / ASR / Audio Captioning ---
openai-whisper>=20230314
torchaudio==2.9.1
torchcodec>=0.0.7
librosa>=0.10.0
soundfile>=0.12.1
ffmpeg_downloader>=0.3.0
ffmpeg-python>=0.1.20

# Noise reduction
noisereduce>=2.0.1

# Note: ffmpeg must be installed on your system
# For Windows, run: winget install ffmpeg